--- roles/container-engine/cri-o/molecule/default/molecule.yml
@@ -1,49 +1,0 @@
----
-driver:
-  name: vagrant
-  provider:
-    name: libvirt
-lint:
-  name: yamllint
-  options:
-    config-file: ../../../.yamllint
-platforms:
-  - name: ubuntu1804
-    box: generic/ubuntu1804
-    cpus: 2
-    memory: 1024
-    groups:
-      - kube-master
-  - name: centos7
-    box: centos/7
-    cpus: 2
-    memory: 1024
-    groups:
-      - kube-master
-  - name: centos8
-    box: centos/8
-    cpus: 2
-    memory: 1024
-    groups:
-      - kube-master
-  - name: fedora
-    box: fedora/31-cloud-base
-    cpus: 2
-    memory: 1024
-    groups:
-      - kube-master
-provisioner:
-  name: ansible
-  env:
-    ANSIBLE_ROLES_PATH: ../../../../
-  config_options:
-    defaults:
-      callback_whitelist: profile_tasks
-  lint:
-    name: ansible-lint
-    options:
-      c: ../../../.ansible-lint
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
