--- roles/container-engine/cri-o/molecule/default/playbook.yml
@@ -1,7 +1,0 @@
----
-- name: Converge
-  hosts: all
-  become: true
-  roles:
-    - role: kubespray-defaults
-    - role: cri-o
