--- roles/container-engine/cri-o/molecule/default/prepare.yml
@@ -1,6 +1,0 @@
----
-- name: Prepare
-  hosts: all
-  gather_facts: False
-  roles:
-    - role: bootstrap-os
