+++ roles/container-engine/cri-o/molecule/fedora/molecule.yml
@@ -1,0 +1,54 @@
+---
+# SPDX-license-identifier: Apache-2.0
+##############################################################################
+# Copyright (c) 2020
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Apache License, Version 2.0
+# which accompanies this distribution, and is available at
+# http://www.apache.org/licenses/LICENSE-2.0
+##############################################################################
+
+driver:
+  name: vagrant
+  provider:
+    name: libvirt
+lint:
+  name: yamllint
+  options:
+    config-file: ../../../.yamllint
+platforms:
+  - name: fedora30
+    box: generic/fedora30
+    cpus: 1
+    memory: 1024
+    groups:
+      - kube-master
+  - name: fedora31
+    box: generic/fedora31
+    cpus: 1
+    memory: 1024
+    groups:
+      - kube-master
+provisioner:
+  name: ansible
+  playbooks:
+    converge: ../resources/converge.yml
+    prepare: ../resources/prepare.yml
+  env:
+    ANSIBLE_ROLES_PATH: ../../../../
+  config_options:
+    defaults:
+      callback_whitelist: profile_tasks
+  lint:
+    name: ansible-lint
+    options:
+      c: ../../../.ansible-lint
+  inventory:
+    host_vars:
+      fedora30:
+        ansible_python_interpreter: /usr/bin/python3
+verifier:
+  name: testinfra
+  directory: ../resources/tests/
+  lint:
+    name: flake8
