+++ roles/container-engine/cri-o/vars/debian.yml
@@ -5,3 +5,9 @@ crio_packages:
   - runc
 
 crio_runc_path: /usr/sbin/runc
+
+crio_kubernetes_version_matrix:
+  "1.18": "1.17"
+  "1.17": "1.17"
+  "1.16": "1.16"
+  "1.15": "1.15"
