+++ roles/container-engine/cri-o/vars/fedora.yml
@@ -4,3 +4,10 @@ crio_packages:
   - cri-tools
 
 crio_conmon: /usr/libexec/crio/conmon
+
+crio_kubernetes_version_matrix:
+  "1.18": "1.17"
+  "1.17": "1.17"
+  "1.16": "1.14"
+  "1.15": "1.14"
+  "1.13": "1.14"
